# api/openapi.yaml
openapi: 3.1.0
info:
    title: 'Voting API'
    version: 1.0.0
    description: 'Secure voting system API — Express, TypeScript, Prisma. Auto-generated from JSDoc comments.'
servers:
    -
        url: 'http://localhost:4000'
        description: 'Local development'
    -
        url: 'https://api.yourdomain.com'
        description: Production
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: 'Paste an **access token** here.'
security:
    -
        bearerAuth: []
...
paths:
...
    /api/voting/request-otp:
        post:
            tags:
                - Voting
            summary: 'Request OTP for voting'
            description: 'Request a one-time password to authenticate a voter for voting'
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - phone
                            properties:
                                phone:
                                    type: string
                                    description: 'Voter''s phone number'
            responses:
                '200':
                    description: 'OTP sent successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string, example: 'OTP sent successfully'}
                                    phone: {type: string}
                '400':
                    description: 'Validation error'
                '404':
                    description: 'Voter not found'
                '409':
                    description: 'Voter has already voted'
                '429':
                    description: 'Rate limit exceeded'
    /api/voting/verify-otp:
        post:
            tags:
                - Voting
            summary: 'Verify OTP for voting'
            description: 'Verify the one-time password sent to the voter via SMS'
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - phone
                                - code
                            properties:
                                phone:
                                    type: string
                                    description: 'Voter''s phone number in international format'
                                    example: '+233240000000'
                                code:
                                    type: string
                                    description: '6-digit OTP code'
                                    example: '123456'
            responses:
                '200':
                    description: 'Authentication successful'
                '400':
                    description: 'Validation error or invalid/expired OTP'
                '404':
                    description: 'Voter not found'
    /api/voting/election:
        get:
            tags:
                - Voting
            summary: 'Get current election and voting status'
            description: 'Get the current active election with positions and candidates, plus voter''s voting status'
            security:
                -
                    bearerAuth: []
            responses:
                '200':
                    description: 'Current election retrieved successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    election: {type: object, properties: {id: {type: string, format: uuid}, title: {type: string}, description: {type: string}, positions: {type: array, items: {type: object, properties: {id: {type: string}, name: {type: string}, order: {type: number}, candidates: {type: array, items: {type: object, properties: {id: {type: string}, fullName: {type: string}, classYearGroup: {type: string}, photoUrl: {type: string}, bio: {type: string}}}}}}}}}
                                    votingStatus: {type: object, properties: {hasVoted: {type: boolean}, votedPositions: {type: array, items: {type: string}}, canVote: {type: boolean}}}
                '401':
                    description: 'Unauthorized - Authentication required'
                '404':
                    description: 'No active election found'
    /api/voting/cast:
        post:
            tags:
                - Voting
            summary: 'Cast votes for all positions'
            description: 'Submit votes for all election positions at once'
            security:
                -
                    bearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            required:
                                - votes
                            properties:
                                votes:
                                    type: array
                                    items: {type: object, required: [positionId], properties: {positionId: {type: string, format: uuid, description: 'Position ID to vote for'}, candidateId: {type: string, format: uuid, nullable: true, description: 'Candidate ID (null for abstain)'}}}
            responses:
                '200':
                    description: 'Votes cast successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message: {type: string}
                                    ballotId: {type: string, format: uuid}
                                    votesCount: {type: number}
                                    timestamp: {type: string, format: date-time}
                '400':
                    description: 'Validation error'
                '401':
                    description: 'Unauthorized - Authentication required'
                '404':
                    description: 'No active election found'
                '409':
                    description: 'Voter has already voted'
                '429':
                    description: 'Rate limit exceeded'